+++ packages/client/components/MentionDropdown.tsx
@@ -60,7 +60,7 @@ export default forwardRef(
       <div className='border-rad z-10 rounded-md bg-white py-1 shadow-lg outline-hidden in-data-[placement="bottom-start"]:animate-slide-down in-data-[placement="top-start"]:animate-slide-up'>
         {items.map((item, idx) => (
           <div
-            data-highlighted={idx === selectedIndex}
+            data-highlighted={idx === selectedIndex ? '' : undefined}
             className={
               'flex w-full cursor-pointer items-center rounded-md px-4 py-1 text-sm leading-8 text-slate-700 outline-hidden hover:bg-slate-200! hover:text-slate-900 focus:bg-slate-200 data-highlighted:bg-slate-100 data-highlighted:text-slate-900'
             }