+++ packages/client/modules/outcomeCard/components/OutcomeCardAssignMenu/TaskFooterTeamAssigneeMenu.tsx
@@ -142,7 +142,8 @@ const TaskFooterTeamAssigneeMenu = (props: Props) => {
           await atmosphere.fetchQuery<TaskFooterTeamAssigneeMenu_viewerIntegrationsQuery>(query, {
             teamId: nextTeam.id
           })
-        const {github, atlassian} = result?.viewer?.teamMember?.integrations ?? {}
+        const safeRes = result instanceof Error ? undefined : result
+        const {github, atlassian} = safeRes?.viewer?.teamMember?.integrations ?? {}
 
         if ((isGitHubTask && !github?.isActive) || (isJiraTask && !atlassian?.isActive)) {
           // viewer is not integrated, now we have these options: