+++ packages/client/subscriptions/NotificationSubscription.ts
@@ -24,6 +24,10 @@ import {
   removeOrgUserNotificationOnNext,
   removeOrgUserNotificationUpdater
 } from '../mutations/RemoveOrgUserMutation'
+import {
+  removeOrgUsersNotificationOnNext,
+  removeOrgUsersNotificationUpdater
+} from '../mutations/RemoveOrgUsersMutation'
 import handleAddNotifications from '../mutations/handlers/handleAddNotifications'
 import {popNotificationToastOnNext} from '../mutations/toasts/popNotificationToast'
 import {updateNotificationToastOnNext} from '../mutations/toasts/updateNotificationToast'
@@ -114,6 +118,9 @@ const subscription = graphql`
       }
       RemoveOrgUserPayload {
         ...RemoveOrgUserMutation_notification @relay(mask: false)
+      }
+      RemoveOrgUsersSuccess {
+        ...RemoveOrgUsersMutation_notification @relay(mask: false)
       }
       InvalidateSessionsPayload {
         ...InvalidateSessionsMutation_notification @relay(mask: false)
@@ -284,6 +291,7 @@ const updateHandlers = {
   InviteToTeamPayload: inviteToTeamNotificationUpdater,
   MeetingStageTimeLimitPayload: meetingStageTimeLimitUpdater,
   RemoveOrgUserPayload: removeOrgUserNotificationUpdater,
+  RemoveOrgUsersSuccess: removeOrgUsersNotificationUpdater,
   StripeFailPaymentPayload: stripeFailPaymentNotificationUpdater,
   ArchiveTimelineEventSuccess: archiveTimelineEventNotificationUpdater
 } as const
@@ -293,6 +301,7 @@ const onNextHandlers = {
   CreateTaskPayload: createTaskNotificationOnNext,
   InviteToTeamPayload: inviteToTeamNotificationOnNext,
   RemoveOrgUserPayload: removeOrgUserNotificationOnNext,
+  RemoveOrgUsersSuccess: removeOrgUsersNotificationOnNext,
   StripeFailPaymentPayload: stripeFailPaymentNotificationOnNext,
   MeetingStageTimeLimitPayload: meetingStageTimeLimitOnNext,
   InvalidateSessionsPayload: invalidateSessionsNotificationOnNext,