+++ packages/server/graphql/public/types/UpgradeToTeamTierSuccess.ts
@@ -4,7 +4,6 @@ import {UpgradeToTeamTierSuccessResolvers} from '../../private/resolverTypes'
 export type UpgradeToTeamTierSuccessSource = {
   orgId: string
   teamIds: string[]
-  meetingIds: string[]
 }
 
 const UpgradeToTeamTierSuccess: UpgradeToTeamTierSuccessResolvers = {
@@ -14,10 +13,6 @@ const UpgradeToTeamTierSuccess: UpgradeToTeamTierSuccessResolvers = {
   teams: async ({teamIds}, _args, {dataLoader}) => {
     const teams = await dataLoader.get('teams').loadMany(teamIds)
     return teams.filter(isValid)
-  },
-  meetings: async ({meetingIds}, _args, {dataLoader}) => {
-    const meetings = await dataLoader.get('newMeetings').loadMany(meetingIds)
-    return meetings.filter(isValid)
   }
 }