+++ packages/server/socketHandlers/handleClose.ts
@@ -1,18 +1,12 @@
-import {WebSocket} from 'uWebSockets.js'
-import sendToSentry from '../utils/sendToSentry'
+import {type WebSocketBehavior} from 'uWebSockets.js'
 import handleDisconnect from './handleDisconnect'
 import {SocketUserData} from './handleOpen'
 
-const handleClose = (ws: WebSocket<SocketUserData>) => {
+const handleClose: WebSocketBehavior<SocketUserData>['close'] = (ws) => {
   const userData = ws.getUserData()
   const {connectionContext} = userData
   if (!connectionContext) return
   userData.done = true
-  try {
-    handleDisconnect(connectionContext)
-  } catch (e) {
-    const error = e instanceof Error ? e : new Error('handleDisconnect failed')
-    sendToSentry(error)
-  }
+  handleDisconnect(connectionContext)
 }
 export default handleClose